Refactor: 사용하지 않는 클래스 파일 삭제
